+++ src/main/java/com/wakeUpTogetUp/togetUp/infra/aws/s3/FileService.java
@@ -5,7 +5,7 @@ import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.wakeUpTogetUp.togetUp.infra.aws.s3.model.CustomFile;
-import com.wakeUpTogetUp.togetUp.utils.FileUtil;
+import com.wakeUpTogetUp.togetUp.utils.file.FileUtil;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
