+++ src/main/java/com/wakeUpTogetUp/togetUp/utils/image/ImageDrawer.java
@@ -6,7 +6,7 @@ import com.wakeUpTogetUp.togetUp.api.mission.model.Coord;
 import com.wakeUpTogetUp.togetUp.common.Status;
 import com.wakeUpTogetUp.togetUp.exception.BaseException;
 import com.wakeUpTogetUp.togetUp.infra.aws.s3.model.ImageContentType;
-import com.wakeUpTogetUp.togetUp.utils.FileUtil;
+import com.wakeUpTogetUp.togetUp.utils.file.FileUtil;
 import java.awt.BasicStroke;
 import java.awt.Color;
 import java.awt.Font;
